requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
clickhouse-connect==0.7.12
clickhouse-driver==0.2.8
dnspython==2.6.1
docker==7.1.0
Flask==3.0.3
greenlet==3.0.3
grpcio==1.62.1
grpcio-tools==1.62.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
lz4==4.3.3
MarkupSafe==2.1.5
motor==3.4.0
protobuf==4.25.3
protobuf-to-dict==0.1.0
pymongo==4.6.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
SQLAlchemy==2.0.30
testcontainers==4.5.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
zstandard==0.22.0
